pyproject.toml
[project]
name = "deeporigin"
description = "Deep Origin Python client"
authors = [{ name = "Deep Origin", email = "support@deeporigin.com" }]
readme = "README.md"
classifiers = [
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities",
]
requires-python = ">=3.10"
dependencies = [
    "requests",
    "termcolor",
    "pandas",
    "beartype",
    "tabulate",
    "httpx",
    "humanize",
    "packaging",
    "tqdm",
    "pyjwt",
    "cryptography",
    "python-box",
    "do-sdk-platform==6.3.0",
    "urllib3>=2.0.0",
    "more-itertools",
    "nest_asyncio",
    "asyncio",
    "jinja2",
    "aiohttp",
    "jsonschema",
]
dynamic = ["version"]

[project.optional-dependencies]
lint = ["ruff"]
test = [
    "pytest",
    "parameterized",
    "coverage",
    "interrogate",
    "pytest-xdist",
    "pytest-markdown-docs",
    "pytest-dependency",
    "fastapi",
    "uvicorn[standard]",
]
dev = ["ipykernel", "jupyter-black"]
docs = [
    "mkdocs",
    "mkdocs-material-extensions",
    "mkdocs-material>=8.2.0",
    "mkdocstrings",
    "mkdocstrings-python",
    "black",
]
plots = ["bokeh"]
tools = [
    "rdkit",
    "deeporigin-molstar==0.1.3",
    "biotite",
    "biopython",
    "rcsb-api>=1.1.4",
]

[tool.setuptools]
package-dir = { "deeporigin" = "src" }
include-package-data = true

[tool.setuptools.dynamic]
version = { file = ["src/VERSION"] }

[project.scripts]
deeporigin = "deeporigin.cli:main"

[build-system]
requires = ["setuptools >= 40.9.0"]
build-backend = "setuptools.build_meta"


[tool.ruff]
include = ["**/src/**/*.py", "**/tests/**/*.py"]
# Exclude file-service auto-gen code for now
exclude = ["src/files/file_service"]
fix = true

[tool.ruff.lint]
ignore = ["E203", "E266", "E501"]
select = ["E4", "E7", "E9", "F", "B", "Q", "F401", "E711", "F841", "F821", "I"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
case-sensitive = false
known-first-party = ["deeporigin"]
known-third-party = ["numpy", "pandas", "rdkit", "beartype"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
force-sort-within-sections = true

[tool.distutils.bdist_wheel]
universal = true

[tool.coverage.run]
omit = ["src/data_hub/client.py", "tests/*.py"]


[tool.pytest.ini_options]
norecursedirs = ["notebooks", "scratch-notebooks"]

[tool.interrogate]
fail-under = 100
# Skip files module coverage for now as it has lots of auto-gen code
exclude = [
    "setup.py",
    "docs",
    "build",
    "tests",
    "variables",
    "conftest.py",
    "src/files",
    "notebooks/*",
    "scratch-notebooks/*",
    "scripts/*",
]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-property-decorators = true
ignore-nested-functions = true
ignore-nested-classes = true
ignore-setters = false
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
color = true
